refactor(invoices): move IIssuedInvoiceRepository and IssuedInvoiceFilters to Persistence.Invoices
Move interface and filter model from Application.Contracts to Persistence.Invoices
to establish correct layer boundaries. Update all consumers with the new namespace.
