+++ backend/src/Anela.Heblo.Application/Features/Invoices/Infrastructure/InvoiceConsumptionSourceAdapter.cs
@@ -1,4 +1,5 @@
 using Anela.Heblo.Application.Features.Invoices.Contracts;
+using Anela.Heblo.Persistence.Invoices;
 using Anela.Heblo.Application.Features.PackingMaterials.Contracts;
 
 namespace Anela.Heblo.Application.Features.Invoices.Infrastructure;
