+++ backend/src/Anela.Heblo.Application/Features/Invoices/Infrastructure/IssuedInvoiceRepository.cs
@@ -1,4 +1,5 @@
 using Anela.Heblo.Application.Features.Invoices.Contracts;
+using Anela.Heblo.Persistence.Invoices;
 using Anela.Heblo.Xcc.Persistance;
 using Anela.Heblo.Domain.Features.Invoices;
 using Anela.Heblo.Persistence;
