+++ backend/src/Anela.Heblo.Application/Features/Invoices/Services/InvoiceImportService.cs
@@ -1,6 +1,7 @@
 using System.ComponentModel;
 using System.Text.Json;
 using Anela.Heblo.Application.Features.Invoices.Contracts;
+using Anela.Heblo.Persistence.Invoices;
 using Anela.Heblo.Application.Features.Invoices.Infrastructure;
 using Anela.Heblo.Domain.Features.Invoices;
 using AutoMapper;
