+++ backend/test/Anela.Heblo.Tests/Features/Invoices/InvoiceImportServiceTests.cs
@@ -1,5 +1,6 @@
 using System.ComponentModel;
 using Anela.Heblo.Application.Features.Invoices.Contracts;
+using Anela.Heblo.Persistence.Invoices;
 using Anela.Heblo.Application.Features.Invoices.Infrastructure;
 using Anela.Heblo.Application.Features.Invoices.Services;
 using Anela.Heblo.Domain.Features.Invoices;
